Fixed all Known Pause menu Bugs
Separated out the pause menu methods into its own script

Streamlimed the Pause menu methods
